Move PlansFeaturesRemote to a separate file
